(1) fix an minor issue wherein e.g. A1D_Get was called after A1D_Get_memcpy for local ops (2) added all A1U_*_memcpy ops in the appropriate ADL functions
it is almost trivial, but i'm not sure if it is better test for pointer nullity or rank=me in the conditional freeing of the NB handle in the "noncontig over contig" implementations.  i use the latter for logical symmetry right now but maybe that's less safe since my_rank is a global variable whereas the handle is local.

git-svn-id: https://svn.mcs.anl.gov/repos/ospri@431 4a0c8edc-5189-0410-a5da-cd986011a6a8